+++ CHANGELOG.rst
@@ -35,6 +35,7 @@ Version 2021-dev
 -  Bump required CMake version to 3.12 (#301)
 -  refactored getenv() into separate function (#302)
 -  use unique_ptr in factory (#303)
+-  add option to print man pages in rst format (#312)
 
 Version 1.6.3 (released 09.12.20)
 =================================